Login and logout functional. ready to merge to develop
