+++ src/ProtectedRoute.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { Route, Redirect, useLocation, useHistory } from 'react-router-dom';
+import { Route, Redirect } from 'react-router-dom';
 import { useAuth } from './contexts/AuthContext';
 
 const ProtectedRoute = ({ children, ...rest }) => {
